パネル制御の変更
