+++ Assets/Scripts/Scenes/Lobby/RoomSettingPanel/RoomSetting_1.cs
@@ -10,12 +10,12 @@ namespace Scenes.Lobby.RoomSettingPanel
 
         public void ToRoomSetting_2()
         {
-            _uiManagerCs.ChangePanel(1);
+            _uiManagerCs.ChangePanel(1); //セッション作成
         }
 
-        public void ToRoomSetting_5()
+        public void ToRoomSetting_3()
         {
-            _uiManagerCs.ChangePanel(4);
+            _uiManagerCs.ChangePanel(2); //セッション参加
         }
     }
 }
