+++ Assets/Scripts/Scenes/Lobby/RoomSettingPanel/RoomSetting_4.cs
@@ -1,54 +1,36 @@
+using System;
 using System.Collections;
 using System.Collections.Generic;
+using System.Linq;
 using UnityEngine;
-using UniRx;
-using UniRx.Triggers;
 using TMPro;
 using Common.Network;
+using Fusion;
 
 namespace Scenes.Lobby.RoomSettingPanel
 {
     public class RoomSetting_4 : MonoBehaviour
     {
-        [SerializeField] private TMP_Text _sessionName;
-        [SerializeField] private List<TMP_Text> _playerNames;
+        [SerializeField] private TMP_InputField _sessionName;
+        [SerializeField] private int _sessionStartSceneIndex = 0;
 
-        private void Start()
+        public async void JoinPrivateSession()
         {
-            this.FixedUpdateAsObservable(). //セッション名の更新
-                Subscribe(_ => UpdateSessionName());
+            //セッションが存在するかを判定
+            string sessionName = _sessionName.text;
+            var sessionInfo = RunnerManager.Instance.SessionInfoList.FirstOrDefault(x => x.Name == sessionName);
+            if (sessionInfo == null) return;
 
-            this.FixedUpdateAsObservable(). //プレイヤーリストの更新
-                Subscribe(_ => UpdatePlayerList());
-        }
-
-        /// <summary>
-        /// セッション名の更新
-        /// </summary>
-        private void UpdateSessionName()
-        {
-            string sessionName = RunnerManager.Runner.SessionInfo.Name;
-            _sessionName.text = "ルームID：" + sessionName;
-        }
-
-        /// <summary>
-        /// プレイヤーリストの更新
-        /// </summary>
-        private void UpdatePlayerList()
-        {
-            var playerInfos = FindObjectsOfType<PlayerInfo>();
-
-            for (int i = 0; i < _playerNames.Count; i++)
+            var args = new StartGameArgs()
             {
-                if (i < playerInfos.Length)
-                {
-                    _playerNames[i].text = playerInfos[i].userName;
-                }
-                else
-                {
-                    _playerNames[i].text = "メンバーがいません";
-                }
-            }
+                GameMode = GameMode.Client, //セッション内権限
+                Scene = SceneRef.FromIndex(_sessionStartSceneIndex), //セッション開始シーン
+                SceneManager = RunnerManager.Runner.GetComponent<NetworkSceneManagerDefault>(), //Fusion用のシーン遷移を管理するコンポーネント
+                SessionName = sessionName, //セッション名
+                ConnectionToken = Guid.NewGuid().ToByteArray(), //プレイヤーの接続トークン
+            };
+
+            var result = await RunnerManager.Runner.StartGame(args); //セッション開始
         }
     }
 }
