--- Assets/Scripts/Scenes/Lobby/RoomSettingPanel/RoomSetting_6.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: fd14394ff0911f14f82e80797d2e0381
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
