--- Assets/Scripts/Scenes/Lobby/RoomSettingPanel/RoomSettion_5.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 86de992529e611a4da3b4a8f5dadef8b
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
